# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/webhook/github": {
      "post": {
        "operationId": "WebhookController_handleGithubWebhook",
        "parameters": [
          {
            "name": "x-hub-signature-256",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "tags": ["Webhook"]
      }
    },
...
  },
  "info": {
    "title": "THESIS",
    "description": "The API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
...
  },
  "security": [{ "bearer": [] }]
}